Add prefix for StatementEvent
