+++ externals/kyuubi-spark-sql-engine/src/main/scala/org/apache/kyuubi/engine/spark/events/SparkStatementEvent.scala
@@ -15,52 +15,36 @@
  * limitations under the License.
  */
 
-package org.apache.kyuubi.events
+package org.apache.kyuubi.engine.spark.events
+
+import org.apache.spark.sql.Encoders
+import org.apache.spark.sql.types.StructType
 
 import org.apache.kyuubi.Utils
-import org.apache.kyuubi.operation.ExecuteStatement
-import org.apache.kyuubi.operation.OperationState.OperationState
 
 /**
- *
- * @param user: who connect to kyuubi server
  * @param statementId: the identifier of operationHandler
  * @param statement: the sql that you execute
- * @param remoteIp: the ip of user
+ * @param appId: application id a.k.a, the unique id for engine
  * @param sessionId: the identifier of a session
  * @param createTime: the create time of this statement
  * @param state: store each state that the sql has
  * @param stateTime: the time that the sql's state change
+ * @param queryExecution: contains logicPlan and physicalPlan
  * @param exception: caught exception if have
  */
-case class StatementEvent(
-    user: String,
+case class SparkStatementEvent(
     statementId: String,
     statement: String,
-    remoteIp: String,
+    appId: String,
     sessionId: String,
     createTime: Long,
     var state: String,
     var stateTime: Long,
-    var exception: String = "") extends KyuubiServerEvent {
+    var queryExecution: String = "",
+    var exception: String = "") extends KyuubiSparkEvent {
+
+  override def schema: StructType = Encoders.product[SparkStatementEvent].schema
   override def partitions: Seq[(String, String)] =
     ("day", Utils.getDateFromTimestamp(createTime)) :: Nil
-}
-
-object StatementEvent {
-  def apply(statement: ExecuteStatement,
-      statementId: String,
-      state: OperationState,
-      stateTime: Long): StatementEvent = {
-    val session = statement.getSession
-    new StatementEvent(
-      session.user,
-      statementId,
-      statement.statement,
-      session.ipAddress,
-      session.handle.identifier.toString,
-      stateTime,
-      state.toString,
-      stateTime)
-  }
 }
